+++ CHANGELOG.md
@@ -79,6 +79,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 
+* (ante) [#1062](https://github.com/crypto-org-chain/ethermint/pull/1062) feat(ante): add hash-keyed sender cache to VerifyEthSig
 * (rpc) [#1048](https://github.com/crypto-org-chain/ethermint/pull/1048) feat(rpc): add `debug_getRawTransaction`
 * (rpc) [#1046](https://github.com/crypto-org-chain/ethermint/pull/1046) feat(rpc): add `debug_getRawBlock`
 * (rpc) [#1047](https://github.com/crypto-org-chain/ethermint/pull/1047) feat(rpc): add `debug_getRawHeader`